Consolidate access permissions into one scenario
See:
- https://github.com/zinc-collective/convene/issues/41
- https://github.com/zinc-collective/convene/issues/40
- https://github.com/zinc-collective/convene/issues/59

We realized that the access permissions bits were a bit hard to track
down because they were spread across the entering rooms and locked rooms
feature definitions.

Because we assume people will be asking "who can enter this room?" it
made sense to consolidate the different permutations into a single
scenario in the Entering Rooms feature definition.